cleanup for map component
