+++ src/components/location-finder/Map/index.js
@@ -1,167 +1,1 @@
-import React, { Component, Fragment, useState } from 'react';
-import * as DS from '@nypl/design-system-react-components';
-// Google map
-import { withGoogleMap, withScriptjs, GoogleMap, Marker, InfoWindow } from 'react-google-maps';
-import { compose } from 'recompose';
-const { NEXT_PUBLIC_GOOGLE_MAPS_API } = process.env;
-// Redux
-import { useDispatch, useSelector } from 'react-redux';
-import { setMapInfoWindow } from './../../../redux/actions';
-// Apollo
-import { useQuery } from '@apollo/client';
-import { LocationsQuery as LOCATIONS_QUERY } from './Locations.gql';
-// Hooks
-import useWindowSize from './../../../hooks/useWindowSize';
-// Utils
-import setTermsFilter from './../../../utils/setTermsFilter';
-
-const MapWrapper = compose(withScriptjs, withGoogleMap)(props => {
-  // Redux
-  const {
-    searchQueryGeoLat,
-    searchQueryGeoLng,
-    openNow,
-    searchFilters
-  } = useSelector(state => state.search);
-  const {
-    mapCenter,
-    mapZoom,
-    infoWindowId,
-    infoWindowIsVisible
-  } = useSelector(state => state.map);
-  
-  // Convert the searchFilters to the object format needed by gql.
-  const termIds = setTermsFilter(searchFilters);
-
-  // Apollo
-  const { loading, error, data } = useQuery(
-    LOCATIONS_QUERY, {
-      variables: {
-        openNow,
-        termIds
-      }
-    }
-  );
-
-  if (loading || !data) {
-    return (
-      <div>Loading</div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div>'error while loading locations'</div>
-    );
-  }
-
-  return (
-    <GoogleMap
-      defaultOptions={{mapTypeControl: false}}
-      zoom={mapZoom}
-      center={mapCenter}
-    >
-      {searchQueryGeoLat && searchQueryGeoLng &&
-        <Marker
-          icon={{
-            url: 'https://maps.google.com/mapfiles/ms/icons/green-dot.png',
-            scaledSize: new google.maps.Size(32, 32)
-          }}
-          position={{
-            lat: searchQueryGeoLat,
-            lng: searchQueryGeoLng
-          }}
-        />
-      }
-      {data.allLocations.locations.map(location => {
-        // Binds onClick from Map prop
-        const onClick = props.onClick.bind(this, location);
-
-        return (
-          <Marker
-            key={location.id}
-            onClick={onClick}
-            icon={{
-              url: 'https://maps.google.com/mapfiles/ms/icons/red-dot.png',
-              scaledSize: new google.maps.Size(32, 32)
-            }}
-            position={{
-              lat: location.geoLocation.lat,
-              lng: location.geoLocation.lng
-            }}
-          >
-            {infoWindowIsVisible && infoWindowId === location.id &&
-              <InfoWindow>
-                <div>
-                  <div>
-                    <DS.Link href={location.url}>
-                      {location.name}
-                    </DS.Link>
-                  </div>
-                  <div>{location.address_line1}</div>
-                  <div>{location.locality}, {location.administrative_area}, {location.postal_code}</div>
-                </div>
-              </InfoWindow>
-            }
-          </Marker>
-        )
-      })}
-    </GoogleMap>
-  );
-});
-
-function Map() {
-  // Redux
-  const dispatch = useDispatch();
-
-  const windowSize = useWindowSize();
-
-  function handleClick(location, event) {
-    dispatch(setMapInfoWindow({
-      infoWindowId: location.id,
-      infoWindowIsVisible: true
-    }));
-
-    // Scroll to location on list when map marker is clicked for desktop only.
-    if (windowSize >= 600) {
-      // Set focus
-      document.querySelector(`#lid-${location.id} a`).focus();
-      // Scroll into view.
-      document.getElementById(`lid-${location.id}`).scrollIntoView({
-        alignToTop: false,
-        behavior: 'smooth'
-      });
-    }
-  }
-
-  return (
-    <Fragment>
-      <MapWrapper
-        aria-hidden="true"
-        onClick={handleClick}
-        googleMapURL={`https://maps.googleapis.com/maps/api/js?v=3.exp&libraries=geometry,drawing,places&key=${NEXT_PUBLIC_GOOGLE_MAPS_API}`}
-        loadingElement={<div style={{ height: `100%` }} />}
-        containerElement={<div style={{ height: `500px` }} />}
-        mapElement={<div style={{ height: `100%` }} />}
-      />
-
-      <DS.Link
-        href="#locations-list"
-        className="locations-list-anchor"
-      >
-        Back to List
-        <DS.Icon
-          blockName="more-link"
-          iconRotation="rotate-180"
-          decorative
-          modifiers={[
-            'right'
-          ]}
-          name="arrow"
-        />
-      </DS.Link>
-    </Fragment>
-  );
-}
-
-export default Map;
+export { default } from './Map';
